Fall back to major mode's tab if unable to expand
In modes where tab is associated to something other than indenting,
smart-tab breaks the behavior, unless the user manually disables it
through `smart-tab-disabled-major-modes'.

If no completion is available, make `smart-tab-default' fall back to tab
behavior as defined in major mode or, if none is defined,
`indent-for-tab-command'.

Possibly makes `smart-tab-disabled-major-modes' superfluous.